Parallel: several bugfixs: state merge, suicide fixup, conflict detect, prefetch, fork
This is a complicated patch, to do some fixup

** fix MergeSlotDB
Since copy-on-write is used, transaction will do StateObject deepCopy before it writes the state;
All the dirty state changed will be recorded in this copied one first, the ownership will be
transfered to main StateDB on merge.
It has a potential race condition that the simple ownership transfer may discard other state changes
by other concurrent transactions.
When copy-on-write is used, we should do StateObject merge.

** fix Suicide
Suicide has an address state read operation.
And it also needs do copy-on-write, to avoid damage main StateDB's state object.

** fix conflict detect
If state read is not zero, should do conflict detect with addr state change first.
Do conflict detect even with current slot, if we use copy-on-write and slotDB reuse, same
slot could has race conditon of conflict.

** disable prefetch on slotDB
trie prefetch should be started on main DB on Merge

** Add/Sub zero balance, Set State
These are void operation, optimized to reduce conflict rate.
Simple test show, conflict rate dropped from ~25% -> 12%

**fix a fork on block 15,338,563
It a nonce conflict caused by opcode: opCreate & opCreate2
Generally, the nonce is advanced by 1 for the transaction sender;
But opCreate & opCreate2 will try to create a new contract, the
caller will advance its nonce too.

It makes the nonce conflict detect more complicated: as nonce is a
fundamental part of an account, as long as it has been changed, we mark
the address as StateChanged, any concurrent access to it will be considered
as conflicted.