Parallel: conflict optimize, remove SlotDB reuse, trie prefetch
** optimize conflict for AddBalance(0)
Add balance with 0 did nothing, but it will do an empty() check, and add
a touch event. Add on transaction finalize, the touch event will check if
the StateObject is empty, do empty delete if it is.

This patch is to take the empty check as a state check, if the addr state has
not been changed(create, suicide, empty delete), then empty check is reliable.

** optimize conflict for system address

** some code improvement & lint fixup & refactor for params

** remove reuse SlotDB
Reuse SlotDB was added to reduce copy of StateObject, in order to mitigate
the Go GC problem.
And COW(Copy-On-Write) is used to address the GC problem too. With COW enabled,
reuse can be removed as it has limitted benefits now and add more complexity.

** fix trie prefetch on dispatcher

Trie prefetch will be scheduled on object finalize.
With parallel, we should schedule trie prefetch on dispatcher, since
the TriePrefetcher is not safe for concurrent access and it is created & stopped
on dispatcher routine.

But object.finalize on slot cleared its dirtyStorage, which broken the later trie
prefetch on dispatcher when do MergeSlotDB.